pyproject.toml
[tool.poetry]
name = "face-aging"
version = "0.1.0"
description = ""
authors = ["cgtyunerr <cgtyuner06@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.1"
pydantic = "^2.8.2"
pydantic-settings = "^2.4.0"
pytest = "^8.3.2"
pillow = "^10.4.0"
numpy = "^1.23.5"
uvicorn = "^0.30.5"
mtcnn = "^0.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
